Email validation
Created an EmailResponse object to store the error that occurred and
the emails that could not be sent. Updated the SendEmailsByBatchIdAndCSV
method in the DistributionController to have the proper httpstatuses.
Added a validateEmail method to the EmailService and EmailServiceImpl.
We updated the sendEmailsByCVS method in the DistributionService to
call a method to validate emails and sendEmails. It now returns an
EmailResponse.